Fix config.ts: read settings and not use internal-build-dir by default (#132)
Before this PR and after 0.2.0, the settings window seems to no longer
detect the options set for some unknown reason. It's possible that the
logic used in the 'VsCodeSettings' file has become outdated due to the
latest updates and the use of a very new vscode engine (+1.90.0).
Despite my attempts to fix its behavior, I ultimately gave up. This PR
now reads options according to the latest API. Additionally, the
internal build directory is no longer used by default; it's now
optional.

This PR goes on top:

- #131

Note: If the user changes any option in the Settings panel, they need to
reload the window or reset the vscode for the changes to take effect.