Add a new status bar item to indicate juvix command status and minors (#140)
This PR revisits the typechecking in the background command
(`juvix-mode.typechecking-silent`) It adds a new status bar item
indicating the result of the typechecking routine and reports the
typechecking error, creating VS Code problems. By default, typechecking
works on save, and it also seems to work fine on changes (the
"typecheckOn" option).